+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- ✨ add get_tag functionality(pr [#538])
+
 ## [0.4.40] - 2025-04-22
 
 ### Added
@@ -1291,6 +1297,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#534]: https://github.com/jerus-org/pcu/pull/534
 [#535]: https://github.com/jerus-org/pcu/pull/535
 [#537]: https://github.com/jerus-org/pcu/pull/537
+[#538]: https://github.com/jerus-org/pcu/pull/538
+[Unreleased]: https://github.com/jerus-org/pcu/compare/v0.4.40...HEAD
 [0.4.40]: https://github.com/jerus-org/pcu/compare/v0.4.39...v0.4.40
 [0.4.39]: https://github.com/jerus-org/pcu/compare/v0.4.38...v0.4.39
 [0.4.38]: https://github.com/jerus-org/pcu/compare/v0.4.37...v0.4.38